非同期が行われなくなってしまっていたので、修正。

+++ app/controllers/messages_controller.rb
@@ -3,7 +3,12 @@ class MessagesController < ApplicationController
 
     def destroy
         message = Message.find(params[:id])
-        message.destroy!
-        redirect_back(fallback_location: root_path)
+        if message.user == current_user
+            message.destroy!
+            redirect_back(fallback_location: root_path)
+        else
+            redirect_to room_path(current_user), notice: '他人のメッセージは削除できません。'
+        end
+        
     end
 end
